test/8_MSA_vs_Monolith/Monolith/monolith/requirements.txt
psycopg2-binary
pika
flask
gunicorn==20.1.0